dynamic dispatch for drag constraint, used for dragCopy && proxyId. dragCopy = true && proxyId means that the origEl is updated; dragging is temporally delegated to the proxy.
Whereas with dragCopy = false && proxyId, the proxyEl is expected to be
absolutely positioned, and the proxy's containing block (could be body)
is positioned (relative; absolute) and is also an ancestor of the
draggable's - el -.

Placeholder tests seem to always pass -- need to investigate
YUI UserAction keypress.

APE.getById - deprecated a long while back; finally removed.