Workaround for buggy input type="date" in WebKit.
Webkit type="date" is a text input with tiny spinner arrows and no
datepicker.  Values are entered by the user by typing and optionally
using the spinner but are not validated. Thus, the UI does not req
ISO-8601 fmt. A feature check makes sure that ISO-8601 fmt is required
by setting non-ISO-8601 fmt and asserting that value is reset to "".
If the buggy input is there, then the type is reset. This causes a
problem, however, in that upon page refresh, the input type="date" is
restored and the user input is lost.